[DW-1668] - centralised the logic for retrieving the Process definition
